requirements.txt
commonmark==0.9.1
numpy==1.22.3
Pillow==9.1.0
Pygments==2.12.0
rich==12.4.1
